+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-# Weather Changelog
+# Get Cat Images Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- Added the main commands
+- Added the Copy functionality
+- Added the Open in Browser functionality
+- Added the Favorites functionality
+